requirements.txt
absl-py==1.0.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
chex==0.1.3
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.2
cycler==0.11.0
dill==0.3.4
dm-tree==0.1.7
einops==0.4.1
etils==0.6.0
flatbuffers==1.12
flax==0.5.1
fonttools==4.33.3
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
grpcio==1.46.1
gviz-api==1.10.0
h11==0.13.0
h5py==3.6.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
jmp==0.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
msgpack==1.0.3
networkx==2.8.4
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
optax==0.1.2
outcome==1.2.0
packaging==21.3
Pillow==9.1.1
promise==2.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rich==11.1.0
rsa==4.8
scikit-image==0.19.3
scipy==1.8.0
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-profile==2.8.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow==2.9.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.9.0
tensorflow-io==0.26.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.8.0
termcolor==1.1.0
tifffile==2022.5.4
toolz==0.11.2
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.0
